Refactor module activities and fragments into their own packages.
