+++ app/src/main/java/com/github/talkbacktutorial/activities/modules/adjustreadingcontrols/AdjustReadingControlsActivity.kt
@@ -1,13 +1,11 @@
-package com.github.talkbacktutorial.activities.modules
+package com.github.talkbacktutorial.activities.modules.adjustreadingcontrols
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.commit
 import com.github.talkbacktutorial.R
-import com.github.talkbacktutorial.databinding.ActivityLesson0Binding
-import com.github.talkbacktutorial.databinding.ActivityMainBinding
-import com.github.talkbacktutorial.databinding.ActivityModuleAdjustReadingControlsBinding
+import com.github.talkbacktutorial.databinding.ActivityAdjustReadingControlsModuleBinding
 
 /**
  * Activity for learning how to adjust reading controls.
@@ -20,7 +18,7 @@ class AdjustReadingControlsActivity : AppCompatActivity() {
      * Data binding used to make manipulating the interface easier. Every element on the layout will
      * be exposed through the binding as a member.
      */
-    lateinit var binding: ActivityModuleAdjustReadingControlsBinding
+    lateinit var binding: ActivityAdjustReadingControlsModuleBinding
         private set
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -30,7 +28,7 @@ class AdjustReadingControlsActivity : AppCompatActivity() {
         supportActionBar?.hide()
 
         // bind activity to layout - this makes our binding work
-        this.binding = DataBindingUtil.setContentView(this, R.layout.activity_module_adjust_reading_controls)
+        this.binding = DataBindingUtil.setContentView(this, R.layout.activity_adjust_reading_controls_module)
 
         // ask the fragment manager to add our fragment, replacing the existing frame in the layout
         supportFragmentManager.commit {
