+++ app/src/main/java/com/github/talkbacktutorial/activities/modules/jumpreadingcontrols/JumpReadingControlsPart2Fragment.kt
@@ -1,4 +1,4 @@
-package com.github.talkbacktutorial.activities.modules
+package com.github.talkbacktutorial.activities.modules.jumpreadingcontrols
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,16 +8,16 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.github.talkbacktutorial.R
 import com.github.talkbacktutorial.TextToSpeechEngine
-import com.github.talkbacktutorial.databinding.FragmentJumpReadingControlsPart3Binding
+import com.github.talkbacktutorial.databinding.FragmentJumpReadingControlsPart2Binding
 
 /**
- * Instantiates a UI for user to interact with word/chars/paragraphs/lines mode
+ * Instantiates a UI for user to interact with controls mode
  *
  * @author Joel Yang
  * @see AdjustReadingControlsActivity
  */
-class JumpReadingControlsPart3Fragment : Fragment() {
-    private lateinit var binding: FragmentJumpReadingControlsPart3Binding
+class JumpReadingControlsPart2Fragment : Fragment() {
+    private lateinit var binding: FragmentJumpReadingControlsPart2Binding
     private lateinit var ttsEngine: TextToSpeechEngine
 
     // callback run after this fragment is created for AdjustReadingControlsActivity
@@ -28,7 +28,7 @@ class JumpReadingControlsPart3Fragment : Fragment() {
             Here, we inflate our layout file (basically, turning the XML into a UI) through
             DataBindingUtil, which will provide our layout binding during the inflate process.
          */
-        this.binding = DataBindingUtil.inflate(inflater, R.layout.fragment_jump_reading_controls_part3, container, false)
+        this.binding = DataBindingUtil.inflate(inflater, R.layout.fragment_jump_reading_controls_part2, container, false)
         return binding.root
     }
 
@@ -39,7 +39,7 @@ class JumpReadingControlsPart3Fragment : Fragment() {
 
     private fun speakIntro() {
         val intro = """
-            
+  
         """.trimIndent()
         this.ttsEngine.speakOnInitialisation(intro)
     }
