+++ app/src/main/java/com/github/talkbacktutorial/activities/modules/jumpreadingcontrols/JumpReadingControlsPart3Fragment.kt
@@ -1,4 +1,4 @@
-package com.github.talkbacktutorial.activities.modules
+package com.github.talkbacktutorial.activities.modules.jumpreadingcontrols
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,16 +8,16 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.github.talkbacktutorial.R
 import com.github.talkbacktutorial.TextToSpeechEngine
-import com.github.talkbacktutorial.databinding.FragmentJumpReadingControlsPart1Binding
+import com.github.talkbacktutorial.databinding.FragmentJumpReadingControlsPart3Binding
 
 /**
- * Instantiates a UI for user to interact with header mode
+ * Instantiates a UI for user to interact with word/chars/paragraphs/lines mode
  *
  * @author Joel Yang
  * @see AdjustReadingControlsActivity
  */
-class JumpReadingControlsPart1Fragment : Fragment() {
-    private lateinit var binding: FragmentJumpReadingControlsPart1Binding
+class JumpReadingControlsPart3Fragment : Fragment() {
+    private lateinit var binding: FragmentJumpReadingControlsPart3Binding
     private lateinit var ttsEngine: TextToSpeechEngine
 
     // callback run after this fragment is created for AdjustReadingControlsActivity
@@ -28,18 +28,18 @@ class JumpReadingControlsPart1Fragment : Fragment() {
             Here, we inflate our layout file (basically, turning the XML into a UI) through
             DataBindingUtil, which will provide our layout binding during the inflate process.
          */
-        this.binding = DataBindingUtil.inflate(inflater, R.layout.fragment_jump_reading_controls_part1, container, false)
+        this.binding = DataBindingUtil.inflate(inflater, R.layout.fragment_jump_reading_controls_part3, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         this.speakIntro()
-
     }
 
     private fun speakIntro() {
         val intro = """
+            
         """.trimIndent()
         this.ttsEngine.speakOnInitialisation(intro)
     }
